Add Wizer option to rename functions.
When initializing a module, it is sometimes useful to rename a function
after the module has been modified. For example, we may wish to
substitute an alternate entry point (for WASI, this is `_start()`)
since we have already performed some initialization that the normal
entry point would do.

This PR adds a `-r` option to allow such renaming. For example,

    $ wizer -r _start=my_new_start -o out.wasm in.wasm

will rename the `my_new_start` export in `in.wasm` to `_start` in
`out.wasm`, replacing the old `_start` export. Multiple renamings are
supported at once.

A followup PR will add a simple C++ example that shows how this can be
used to run C++ static constructors and other program initialization
with Wizer, and then invoke `main()` without re-running these
constructors.