Add a fuzz target for Wizer
This fuzz target checks that we get the same result whether we

1. Call the initialization function
2. Call the main function

or

1. Call the initialization function
2. Snapshot with Wizer
3. Instantiate the snapshot
4. Call the instantiated snapshot's main function

When checking that we get the same result, we don't just consider the main
function's results: we also consider memories and globals.

I've been running this locally for a couple days and, while there was an initial
flurry of failures and panics, it's been about a day since it fell over. I think
module linking support is in a good enough place where we can land it and
continue to bug fix in tree as we find new failures.