Do not create "native" sized/aligned data segments
Instead, create minimally sized pages (modulo when it is more size efficient to
merge to nearby segments together due to the size overhead of defining an active
data segment).

The thinking previously was that Wasm engines could recognize when segments were
page sized and aligned and then do virtual memory tricks to lazily initialize
Wasm memories, do copy on write, etc... But they can do this kind of thing by
preprocessing data segments themselves in the module compilation phase (and both
Wasmtime and Lucet do some of this today) and no engine (to my knowledge) will
recognize when the segments are already processed in this manner but won't
preprocess them itself.