Publish binary releases of the `wizer` CLI tool
This adds automation to publish releases for Linux, macOS, and Windows, all 64bit. Same as for Wasmtime, releases are published under the `dev` tag whenever a PR is merged, and whenever a commit is tagged.

This is based on what Wasmtime used to do and what fastly/js-compute-runtime still does. There's quite a bit of duplication of project names and the likes going on, which would be nice to clean up, but I'd expect that to take meaningful amounts of time that I can't currently spend :(