add `initialize_staged` function
This allows the caller to specify two separate components during
pre-initialization: one to instrument (which we call "stage 1"), and the other
to which to apply the snapshot captured after pre-initialization (which we call
"stage 2").  Normally, you'll want these to be the same component, but there are
cases where they might not be: e.g. when the "stage 1" component needs access to WASI
imports during pre-init, but the "stage 2" component does not have those imports.

The second parameter to `initialize_staged` is an optional tuple of the "stage
1" component and a function which maps module indices from "stage 2" to "stage
1".

Signed-off-by: Joel Dice <joel.dice@fermyon.com>