Switch the source-of-truth to Wasmtime (#140)
* Switch the source-of-truth to Wasmtime

This commit is a follow-up to the merging of the Wizer repository
into Wasmtime. At this time I don't have plans for what to do about the
npm package or the old `wizer` CLI, so I've updated this repository
in-place to use the Wasmtime repository upstream. I figure that if other
folks are interested to maintain this then this is a reasonable-enough
starting point and can take it from here. In this commit I deleted all
the benchmarks but kept the tests running.

* Remove fuzzing infrastructure